Fix for EHR Measure table.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@5872 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Misc/ConvertDatabases2.cs
@@ -4113,8 +4113,8 @@ VALUES('MercuryDE','"+POut.String(@"C:\MercuryDE\Temp\")+@"','0','','1','','','1
 					command=@"CREATE TABLE ehrmeasure (
 						EhrMeasureNum bigint NOT NULL auto_increment PRIMARY KEY,
 						MeasureType tinyint NOT NULL,
-						Numerator tinyint NOT NULL,
-						Denominator tinyint NOT NULL
+						Numerator smallint NOT NULL,
+						Denominator smallint NOT NULL
 						) DEFAULT CHARSET=utf8";
 					Db.NonQ(command);
 				}
@@ -4124,8 +4124,8 @@ VALUES('MercuryDE','"+POut.String(@"C:\MercuryDE\Temp\")+@"','0','','1','','','1
 					command=@"CREATE TABLE ehrmeasure (
 						EhrMeasureNum number(20) NOT NULL,
 						MeasureType number(3) NOT NULL,
-						Numerator number(3) NOT NULL,
-						Denominator number(3) NOT NULL,
+						Numerator number(11) NOT NULL,
+						Denominator number(11) NOT NULL,
 						CONSTRAINT ehrmeasure_EhrMeasureNum PRIMARY KEY (EhrMeasureNum)
 						)";
 					Db.NonQ(command);
